// src/Types/DevolucionProducto.tsx
export interface GetDevolucionProducto {
  codigoProducto: string;
  codigo: string;
  cantidadBuena: number;
  estado: string;
  descripcion: string;
}

export interface PostDevolucionProducto {
  codigoProducto: string;
  codigo: string;
  cantidadBuena: number;
  estado: string;
  descripcion: string;
}